// java-components/management-console/src/main/webui/src/services/openapi/models/BuildSummaryDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BuildSummaryDTO = {
    totalBuilds: number;
    successfulBuilds: number;
    contaminatedBuilds: number;
    runningBuilds: number;
    failingBuilds: number;
};